backend/requirements.txt
flask
python-dotenv
requests
pdfkit
jinja2
reportlab
flask-cors